+++ ATMSystem.Data/Entities/Card.cs
@@ -12,7 +12,7 @@ namespace ATMSystem.Data.Entities
         public int Id { get; set; }
         [Required]
         [CreditCard]
-        public int Number { get; set; }
+        public Int64 Number { get; set; }
         [Required]
         public int Password { get; set; }
         [Required]
